+++ packages/yoshi-flow-editor/src/wrappers/settingsWrapping.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import fs from 'fs-extra';
-import { FlowEditorModel, ComponentModel } from './model';
+import { FlowEditorModel, ComponentModel } from '../model';
 
 const settingsWrapperPath =
   'yoshi-flow-editor-runtime/build/SettingsWrapper.js';
